made code platform independant wrt 32/64 bit machines

+++ CodeExamples/SizeOf/sizeOfExpression.c
@@ -5,12 +5,12 @@
 #define OK 0
 
 int main() {
-    printf(" 2    :\t%02d Bits\n",sizeof(2)*8); 
-    printf(" 2l   :\t%02d Bits\n",sizeof(2l)*8); 
-    printf(" 2ll  :\t%02d Bits\n",sizeof(2ll)*8); 
-    printf(" 2.0  :\t%02d Bits\n",sizeof(2.0)*8); 
-    printf(" 2.0f :\t%02d Bits\n",sizeof(2.0f)*8); 
-    printf(" 2.0l :\t%02d Bits\n",sizeof(2.0l)*8); 
+    printf(" 2    :\t%3ld Bits\n", (long unsigned int) sizeof(2)*8); 
+    printf(" 2l   :\t%3ld Bits\n", (long unsigned int) sizeof(2l)*8); 
+    printf(" 2ll  :\t%3ld Bits\n", (long unsigned int) sizeof(2ll)*8); 
+    printf(" 2.0  :\t%3ld Bits\n", (long unsigned int) sizeof(2.0)*8); 
+    printf(" 2.0f :\t%3ld Bits\n", (long unsigned int) sizeof(2.0f)*8); 
+    printf(" 2.0l :\t%3ld Bits\n", (long unsigned int) sizeof(2.0l)*8); 
 
     return OK;
 }
